mejora del sidebar
