+++ src/app.module.ts
@@ -7,7 +7,7 @@ import { GalaxiasModule } from './infraestructure/http/galaxia/galaxias.module';
 import { PlanetasModule } from './infraestructure/http/planeta/planetas.module';
 import { CategoriaModule } from './infraestructure/http/categoria/categorias.module';
 import { EventEmitterModule } from '@nestjs/event-emitter';
-import { MenuSidebarModule } from './infraestructure/menu-sidebar/menuSidebar.module';
+import { MenuModule } from './infraestructure/Menu/menu.module';
 import { AzureModule } from './infraestructure/http/azure/azure.module';
 import { ProfesorModule } from './infraestructure/http/profesor/profesor.module';
 
@@ -21,7 +21,7 @@ import { ProfesorModule } from './infraestructure/http/profesor/profesor.module'
     IdiomaModule,
     ProfesorModule,
     AzureModule,
-    MenuSidebarModule,
+    MenuModule,
   ],
   controllers: [AppController],
   providers: [AppService],
