+++ src/core/repositories/Menus/menu-repository.interface.ts
@@ -1,0 +1,7 @@
+import { MenuEntity } from 'src/core/entities/menu-sidebar/menuSidebar.entity'
+import {MenuSidebar} from '@prisma/client'
+
+export interface MenuSidebarRepository {
+  findByNode(node: string): Promise<MenuEntity[]>
+  findAllActive(): Promise<MenuSidebar[]> 
+}
