--- src/core/repositories/menu-sidebar/menu-repository.interface.ts
@@ -1,6 +1,0 @@
-import {MenuSidebar} from '@prisma/client'
-
-export interface MenuRepository {
-  findByNode(node: string): Promise<MenuSidebar[]>
-}
-
